+++ packages/cli/__tests__/zosfiles/__system__/download/uss/cli.files.download.uss.system.test.ts
@@ -224,7 +224,7 @@ describe("Download USS File", () => {
                 ussname + ".dummy"
             ]);
             expect(response.status).toBe(1);
-            expect(response.stderr.toString()).toContain("File not found.");
+            expect(response.stderr.toString()).toContain("No such file or directory.");
         });
     });
 });