+++ packages/zosfiles/__tests__/__unit__/methods/create/Create.unit.test.ts
@@ -794,10 +794,7 @@ describe("Create data set", () => {
         expect(response.commandResponse).toContain("created successfully");
         expect(mySpy).toHaveBeenCalledWith(dummySession, endpoint, [ZosmfHeaders.ACCEPT_ENCODING], JSON.stringify({
             ...CreateDefaults.DATA_SET.BLANK,
-            ...dsBlankOptions,
-            ...{
-                secondary: 1
-            }
+            ...dsBlankOptions
         }));
     });