+++ zowex/src/test.rs
@@ -18,8 +18,6 @@ use std::time::Duration;
 
 #[cfg(target_family = "unix")]
 use home::home_dir;
-#[cfg(target_family = "windows")]
-use whoami::username;
 
 // Zowe daemon executable modules
 use crate::defs::*;
@@ -114,26 +112,50 @@ fn unit_test_util_get_socket_string() {
 
 #[test]
 fn unit_test_get_zowe_env() {
-    let env = util_get_zowe_env();
-    assert_eq!(env.get("ZOWE_EDITOR"), None);
+    let environment = util_get_zowe_env();
+    assert_eq!(environment.get("ZOWE_EDITOR"), None);
 
     env::set_var("ZOWE_EDITOR", "nano");
-    let env = util_get_zowe_env();
-    assert_eq!(env.get("ZOWE_EDITOR"), Some(&"nano".to_owned()));
-
+    let environment = util_get_zowe_env();
+    assert_eq!(environment.get("ZOWE_EDITOR"), Some(&"nano".to_owned()));
     env::remove_var("ZOWE_EDITOR");
+
+    env::remove_var("FORCE_COLOR");
+    let environment = util_get_zowe_env();
+    let color = util_terminal_supports_color();
+    assert_eq!(environment.get("FORCE_COLOR"), Some(&color.to_string()));
+
+    env::set_var("FORCE_COLOR", "0");
+    let environment = util_get_zowe_env();
+    assert_eq!(environment.get("FORCE_COLOR"), Some(&"0".to_owned()));
+    env::remove_var("FORCE_COLOR");
+
+    env::set_var("FORCE_COLOR", "1");
+    let environment = util_get_zowe_env();
+    assert_eq!(environment.get("FORCE_COLOR"), Some(&"1".to_owned()));
+    env::remove_var("FORCE_COLOR");
+
+    env::set_var("FORCE_COLOR", "2");
+    let environment = util_get_zowe_env();
+    assert_eq!(environment.get("FORCE_COLOR"), Some(&"2".to_owned()));
+    env::remove_var("FORCE_COLOR");
+
+    env::set_var("FORCE_COLOR", "3");
+    let env = util_get_zowe_env();
+    assert_eq!(env.get("FORCE_COLOR"), Some(&"3".to_owned()));
+    env::remove_var("FORCE_COLOR");
 }
 
-#[test]
+#[tokio::test]
 // test daemon restart
-fn integration_test_restart() {
+async fn integration_test_restart() {
     let njs_zowe_path = util_get_nodejs_zowe_path();
 
     let mut daemon_proc_info = proc_get_daemon_info();
     if daemon_proc_info.is_running {
         println!("--- test_restart: To initializes test, stop a running daemon.");
         let mut restart_cmd_args: Vec<String> = vec![SHUTDOWN_REQUEST.to_string()];
-        if let Err(err_val) = run_daemon_command(&njs_zowe_path, &mut restart_cmd_args) {
+        if let Err(err_val) = run_daemon_command(&njs_zowe_path, &mut restart_cmd_args).await {
             assert_eq!(
                 "Shutdown should have worked", "Shutdown failed",
                 "exit code = {}",
@@ -151,7 +173,7 @@ fn integration_test_restart() {
 
     // now try the restart
     println!("--- test_restart: Run a restart when no daemon is running.");
-    let result = run_restart_command(&njs_zowe_path);
+    let result = run_restart_command(&njs_zowe_path).await;
     assert_eq!(result.unwrap(), 0, "The run_restart_command failed.");
 
     // confirm that the daemon is running
@@ -164,7 +186,7 @@ fn integration_test_restart() {
     let first_daemon_pid = daemon_proc_info.pid;
 
     println!("--- test_restart: Run a restart with a daemon already running.");
-    let result = run_restart_command(&njs_zowe_path);
+    let result = run_restart_command(&njs_zowe_path).await;
     assert_eq!(result.unwrap(), 0, "The run_restart_command failed.");
 
     // confirm that a new and different daemon is running
@@ -183,7 +205,7 @@ fn integration_test_restart() {
     // As a cleanup step, stop the daemon
     println!("--- test_restart: To cleanup, stop the running daemon.");
     let mut restart_cmd_args: Vec<String> = vec![SHUTDOWN_REQUEST.to_string()];
-    if let Err(_err_val) = run_daemon_command(&njs_zowe_path, &mut restart_cmd_args) {
+    if let Err(_err_val) = run_daemon_command(&njs_zowe_path, &mut restart_cmd_args).await {
         assert_eq!("Shutdown should have worked", "Shutdown failed");
     }