+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## `7.16.4`
+
+BugFix: Fixed `secondary` option being specified as `1` on `BLANK` type datasets with the `zowe files create data-set` command [#1595](https://github.com/zowe/zowe-cli/issues/1595)
+
+## `7.16.3`
+
+- BugFix: Updated `imperative` to fix undesired behavior in the `zowe config list` command in certain situations.
+
+## `7.16.2`
+
+- BugFix: Updated `tar` dependency.
+
+## `7.16.1`
+
+- BugFix: Fixed `--range` option ignored on `zowe files view uss-file` command.
+- BugFix: Fixed `--binary` option ignored by commands that upload and download USS directories when ".zosattributes" file is used.
+- BugFix: Fixed `--include-hidden` option ignored by `zowe files upload dir-to-uss` without the `--recursive` option.
+
+## `7.16.0`
+
+- Enhancement: Updated daemon to use `tokio` library instead of unmaintained `named_pipe` library.
+
+## `7.15.0`
+
+- Enhancement: Added the `zowe files copy dsclp` command to copy a dataset from one LPAR to another.  [#1098](https://github.com/zowe/zowe-cli/issues/1098)
+
+## `7.14.1`:
+
+- Enhancement: Re-enabled color in the daemon client [#1379](https://github.com/zowe/zowe-cli/issues/1379)
+- BugFix: Enabled ANSI in Windows based terminals [#1701](https://github.com/zowe/zowe-cli/issues/1701)
+- BugFix: Changed daemon to spawn as its own process [#1241](https://github.com/zowe/zowe-cli/issues/1241) [#1277](https://github.com/zowe/zowe-cli/issues/1277) [#1309](https://github.com/zowe/zowe-cli/issues/1309)
+- BugFix: Updated Imperative to allow for special handling of chalk and coloring in daemon client
+
 ## `7.13.0`
 
 - Enhancement: Updated Imperative to add `--prune` option to `zowe config secure` command. [Imperative #547](https://github.com/zowe/imperative/issues/547)